[Refactor] Init and use per_request_initial_chunk_size_override helper
